refactor: move bash related logic into `BashSession` for cleaner code (#4527)
