+++ openhands/runtime/utils/__init__.py
@@ -1,7 +1,6 @@
-from openhands.runtime.utils.bash import split_bash_commands
 from openhands.runtime.utils.system import (
     display_number_matrix,
     find_available_tcp_port,
 )
 
-__all__ = ['display_number_matrix', 'find_available_tcp_port', 'split_bash_commands']
+__all__ = ['display_number_matrix', 'find_available_tcp_port']
